Install repose from R-Universe if needed
The building of the pkgdown website failed because we changed the
NGCHMDemoData and NGCHMSupportFiles repo to be under
'Additional_repositories'. While this is the thing that people do,
pkgdown does not use this field to install the 'Suggests' packages. So
modifying the vignettes to install from our R-Universe repo if needed.